test(ios): split the service tests where the file went over budget
`PersonalSuggestionServiceCreationTests` reached 159 lines, nine over the Swift
limit, when the prediction cases were added to it. The context tests are their
own subject anyway — what the service does with the word a prefix follows — and
they move to a file of that name.

Nothing caught this: `Scripts/check-swift-loc.sh` runs in the deploy workflow, not
on pull requests, so twelve green PRs sat on top of an over-budget file. Wiring it
into the PR gate is its own change.
